+++ csrc/swizzle.cpp
@@ -73,7 +73,7 @@ std::pair<Val*, Val*> unCyclicShift(Val* x, Val* y, Val* size_x) {
 // Scatter swizzle:
 //   Corresponds to the data layout out of ldmatrix intrinsic.
 //   supported dimensions are : 8x4, 16x4, 32x4
-std::pair<Val*, Val*> Scatter(Val* x, Val* y, int size_x) {
+std::pair<Val*, Val*> Scatter(Val* x, Val* y, int64_t size_x) {
   TORCH_CHECK(
       size_x == 8 || size_x == 16 || size_x == 32,
       "Unsupported Scatter swizzle size");
@@ -82,7 +82,7 @@ std::pair<Val*, Val*> Scatter(Val* x, Val* y, int size_x) {
   return {cpp_div(add(mul(y, size_x_val), x), four), mod(x, four)};
 }
 
-std::pair<Val*, Val*> unScatter(Val* x, Val* y, int size_x) {
+std::pair<Val*, Val*> unScatter(Val* x, Val* y, int64_t size_x) {
   TORCH_CHECK(
       size_x == 8 || size_x == 16 || size_x == 32,
       "Unsupported Scatter swizzle size");
@@ -107,7 +107,7 @@ std::pair<Val*, Val*> dispatchSwizzle(
     case Swizzle2DType::CyclicShift:
       return swizzles::CyclicShift(x, y, maybe_size_x);
     case Swizzle2DType::Scatter:
-      return swizzles::Scatter(x, y, (int)maybe_size_x->evaluateInt());
+      return swizzles::Scatter(x, y, maybe_size_x->evaluateInt());
     default:
       TORCH_INTERNAL_ASSERT(false, "Unsupported swizzle type");
   }
@@ -127,7 +127,7 @@ std::pair<Val*, Val*> dispatchUnSwizzle(
     case Swizzle2DType::CyclicShift:
       return swizzles::unCyclicShift(x, y, maybe_size_x);
     case Swizzle2DType::Scatter:
-      return swizzles::unScatter(x, y, (int)maybe_size_x->evaluateInt());
+      return swizzles::unScatter(x, y, maybe_size_x->evaluateInt());
     default:
       TORCH_INTERNAL_ASSERT(false, "Unsupported swizzle type");
   }
